requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
WTForms==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
openpyxl==3.1.2
Flask-Migrate==4.0.4 